requirements.txt
django-heroku
gunicorn
django
django-environ
flake8
black
coverage
coveralls
pyrebase4
datetime
django-crispy-forms
pyyaml
django-allauth
pillow
model_bakery
django-notifications-hq
